requirements.txt
pytest==9.0.1
requests==2.32.5
pytest-html==4.1.1
pytest-xdist==3.8.0
python-dotenv==1.2.1


